backend/requirements/requirements-base.txt
Django>=4.2.3
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-camel-case==1.4.2
djangorestframework-simplejwt==5.2.2
django-cors-headers==4.2.0
drf-spectacular==0.26.3
python-decouple==3.6
python-dotenv==1.0.0
drf-nested-routers==0.93.4
django-model-utils==4.3.1
django-allauth==0.54.0
dj-rest-auth==4.0.1
drf-extensions==0.7.1
psycopg2==2.9.6
Markdown==3.4.3
Pillow==9.2.0
tzdata
django-lifecycle

# DI
# =====================
dependency-injector==4.41.0

# my private packages
# 環境変数 GIT_TOKEN に github のトークンをセット
# =====================
git+https://cladder-prj:${GIT_TOKEN}@github.com/tamura-up/django-ffm.git
